**pkginfo now includes a 'path' parameter pointing to the path of the current catpkg directory being processed.
BreezyBuild now takes a parameter template_path to allow you to
specify where to look for templates. Default is in the catpkg
directory for the catpkg it is currently generating.

But you can override this by passing:
  template_path=another_breezybuild.template_path

This will allow you to keep all your templates in the current
catpkg. See kit-fixups/editors-kit/curated/app-editors/vim for
it in use to get a feel for how this works.