feat: add initial meson utilities
This patch implements utilities, which allow us to easily gain
information about a Meson project.

It makes use of the `meson builddir` command, which generates a bunch of
JSON files with info about available benchmarks, build options, build
files, dependencies, etc.

This will be particularly useful for autogenning packages, which use the
Meson build system, as it would allow us to figure out the list of
dependencies along with version requirements, the list of available
build options, etc.