fix: escape all capital letters in gosum
This patch ensures all capital letters in a gosum line are escaped to
resolve issues with the golang proxy calling our versions invalid. We
used to perform this escape, but we only did it for the module name and
not for the version, so I've escaped the entire line before splitting it
to ensure nothing goes wrong. Also, we used to do the escaping with a
regex, I've written a simple and easier to understand (imo) escaping
function instead.