Initial work on release.py to include the concept of 'sourced' kits in git master. The concept here is that each sourced kit can have a direct specification of the source as:
  source:
    url: https://....
    branch: ...

This data in the release.yaml allows direct specification of the source repository
since it's messy to create a source collection and repo definition for each sourced
kit (since this is just redundant.) I have also started to work on initialization
of the YAML object model to reflect this.

Currently, I notice that ``merge-model.current_source_def`` (see kit.py, line 740)
is antiquated as sourced kits don't really *have* a source def to track. So this
needs a bit of overhaul so that sourced kits just initialize their source repo
directly and this should allow them to run in parallel nearly all the time.

Then I need to complete generate_sourced_kit(), once we have sane objects to model
the kits. I currently set kit_insides['source'] to a SourceRepository instead of
a SourceCollection in the case of KitKind.SOURCED, but I don't have the proper
attributes set up to store this info and the code is currently assuming this is
a SourceCollection in a number of places.