This is a fix for JSON fetching. With httpx, sometimes escaping was broken (related to backslashes in strings.) It turns out that with httpx, the most reliable way to get proper json back is to use the json() method on the response object directly rather than rolling your own JSON decoding.
This is a temporary fix. I have also added some TODOs as right now,
is_json= is not written to the store, and not checked, so we assume
that if a cached resource is already cached, it's in the format you
are expecting already. This *can* create problems and *should* be
fixed but should not really cause issues right now.