FL-9827: allow official generators to be used outside of kit-fixups:
"doit" now allows the use of official generators such as github-1
to be used outside of kit-fixups. The way this works is that doit
will now detect if it is being run inside kit-fixups or not. If it
is NOT, then it will clone/update a kit-fixups repository to
~/repo_tmp/source-trees/kit-fixups to have access to the official
generators.

New options have been added to "doit":

  --fixups_url and --fixups_branch:

  Can be used to point to a kit-fixups to clone if you want to use
  a non-standard kit-fixups for access to custom generators.

  --fast:

  If you are using "doit" outside of kit-fixups, doing a git pull
  on ~/repo_tmp/source-trees/kit-fixups can take a bit of time.
  This option tells "doit" to use the already-cloned kit-fixups
  as-is without trying to update it by hand first.

To use:

Simply run "doit" in an arbitrary overlay -- it will "just work".
The extra options are there if needed but are not normally required.

I have confirmed this working by using the github-1 generator in
gnome-kit sources for pipewire.