Improve default pattern of `RegexMatcher`
This patch improves the default pattern of `RegexMatcher` by allowing it
to match patchlevels and revisions, in that particular order. While this
regex is far from matching every version that's considered valid by
Portage, it's at least more robust than simply matching digits and dots.

The `RegexMatcher` could be replaced entirely by a Portage version
parser (if we ever end up making one as a replacement for
`packaging.version`, refer to FL-9449).