All this code is actively being tested and may contain bugs.
Attempt a full rewrite of error handling for autogen and fetch tasks.
I eliminated the use of "done callbacks" and instead have the
gather_pending_tasks() function return two lists -- list of results of
successfully-completed tasks, and a list of tasks that threw exceptions.

I did this because the following exception was not causing an abort of
doit:

FetchRequest <metatools.fastpull.spider.FetchRequest object at 0x7f1d28516d60>
[ERROR 19:17:03] Artifact for url https://fastpull-us.funtoo.org/distfiles/dyndns-samples.tbz2 referenced in net-dns/bind net-dns/bind-tools could not be fetched.
    Traceback (most recent call last):
      File "/home/tree/development/funtoo-metatools/funtoo/pkgtools/ebuild.py", line 528, in wrapper
        await self.generate()
      File "/home/tree/development/funtoo-metatools/funtoo/pkgtools/ebuild.py", line 648, in generate
        await self.setup()
      File "/home/tree/development/funtoo-metatools/funtoo/pkgtools/ebuild.py", line 514, in setup
        raise BreezyError(f"Artifact for url {artifact.url} referenced in {artifact.catpkgs} could not be fetched.")
    dyne.org.funtoo.metatools.pkgtools.ebuild.BreezyError: Artifact for url https://fastpull-us.funtoo.org/distfiles/dyndns-samples.tbz2 referenced in net-dns/bind net-dns/bind-tools could not be fetched.

The use of AUTOGEN_FAILURES did not appear to work, even though in theory
we were appending to this list. Rather than use appropriate thread locking,
it was better to move to a more robust method.

The best action was to not use this global variable but instead handle errors
via the return value of gather_pending_tasks() for those who use this
function to wait for a bunch of tasks to complete.