Implement cargo support via dynamic archives
This patch adds a new `add_crates_bundle` function which uses metatools'
dynamic archives to create a bundle of all crates that are necessary to
build a given Rust project.

This is the equivalent of the newly added "add_gosum_bundle" function
but for Rust. It was used as a reference for the implementation of this
patch and follows a similar API for consistency.